requirements.txt
torch==2.1.2
lightning==2.1.3
torchmetrics==1.3.0
monai==1.3.0
argparse==1.4.0
hydra-core==1.3.2
omegaconf==2.3.0
tqdm==4.66.1
nibabel==5.2.0
rootutils==1.0.7
scikit-learn==1.4.0
scipy==1.11.4

matplotlib==3.8.2
matplotlib-inline==0.1.6
seaborn==0.13.1
